Поменял работу с проектами и организациями

+++ backend/src/main/java/ru/urfu/backend/controller/ProjectController.java
@@ -4,21 +4,22 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.urfu.backend.PathsConstants;
 import ru.urfu.backend.dto.CreatedResponse;
 import ru.urfu.backend.dto.DeletedResponse;
 import ru.urfu.backend.dto.LeftResponse;
-import ru.urfu.backend.dto.invite.InviteRequest;
-import ru.urfu.backend.dto.invite.InviteResponse;
+import ru.urfu.backend.dto.invite.ProjectInviteRequest;
+import ru.urfu.backend.dto.invite.ProjectInviteResponse;
 import ru.urfu.backend.dto.project.*;
 import ru.urfu.backend.dto.tasks.TaskCreateRequest;
 import ru.urfu.backend.exception.customEx.UserNotFoundException;
+import ru.urfu.backend.mapper.ProjectInviteMapper;
 import ru.urfu.backend.mapper.ProjectMapper;
 import ru.urfu.backend.mapper.TaskMapper;
 import ru.urfu.backend.model.*;
+import ru.urfu.backend.model.enums.ProjectMemberRole;
 import ru.urfu.backend.service.*;
 
 import java.util.ArrayList;
@@ -37,60 +38,136 @@ public class ProjectController {
     private final AuthService authService;
     private final TaskService taskService;
     private final TaskMapper taskMapper;
+    private final ProjectInviteService projectInviteService;
+    private final ProjectInviteMapper projectInviteMapper;
 
     @Autowired
     public ProjectController(
-            ProjectMapper projectMapper, ProjectService projectService, OrganizationService organizationService, AuthService authService, TaskService taskService, TaskMapper taskMapper) {
+            ProjectMapper projectMapper,
+            ProjectService projectService,
+            OrganizationService organizationService,
+            AuthService authService,
+            TaskService taskService,
+            TaskMapper taskMapper,
+            ProjectInviteService projectInviteService,
+            ProjectInviteMapper projectInviteMapper) {
         this.projectMapper = projectMapper;
         this.projectService = projectService;
         this.organizationService = organizationService;
         this.authService = authService;
         this.taskService = taskService;
         this.taskMapper = taskMapper;
+        this.projectInviteService = projectInviteService;
+        this.projectInviteMapper = projectInviteMapper;
     }
 
-    //FIXME: Пагинация
-    @GetMapping
-    public List<ProjectListItemDto> getProjects() throws UserNotFoundException {
+    @Operation(description = "Получение проектов текущего пользователя")
+    @GetMapping("/my")
+    public List<ProjectListItemDto> getProjects(
+            @RequestParam(required = false) String name,
+            @RequestParam(required = false) Boolean isPrivate,
+            @RequestParam(required = false) Long organizationId,
+            @RequestParam(required = false) ProjectMemberRole membership
+    ) throws UserNotFoundException {
         User user = authService.getAuthenticatedUser();
-        Set<UserProject> projects = user.getProjects();
-        List<ProjectListItemDto> projectListItemDtos = new ArrayList<>();
-        for (UserProject userProject : projects) {
-            ProjectListItemDto projectListItemDto = projectMapper.mapToProjectListItemDto(userProject);
-            projectListItemDtos.add(projectListItemDto);
+        List<UserProject> userProjects = new ArrayList<>();
+        for (UserProject userProject : user.getProjects()) {
+            Project project = userProject.getProject();
+            if((name != null && name.equals(project.getTitle()))
+                || (isPrivate != null && !isPrivate.equals(project.getIsPrivate()))
+                || (organizationId != null && !organizationId.equals(project.getOrganization().getId()))
+                || (membership != null && !membership.equals(userProject.getProjectMemberRole()))){
+                continue;
+            }
+            userProjects.add(userProject);
         }
-        return projectListItemDtos;
+        return projectMapper.mapToProjectListItemDtos(userProjects);
     }
 
-    @Operation(description = "Создание проекта для организации по её id")
+    @Operation(description = "Получение участников проекта по id")
+    @GetMapping("/{projectId}/participants")
+    public List<ProjectParticipantDto> getParticipants(
+            @PathVariable("projectId") Long projectId,
+            @RequestParam(required = false) String fullName,
+            @RequestParam(required = false) String login,
+            @RequestParam(required = false) ProjectMemberRole role
+    ){
+        Project project = projectService.getById(projectId);
+        List<UserProject> userProjects = new ArrayList<>();
+        for(UserProject userProject : project.getUsers()){
+            User user = userProject.getUser();
+            if((fullName != null && !fullName.equals(user.getFullName()))
+                || (login != null && !login.equals(user.getLogin()))
+                || (role != null && !role.equals(userProject.getProjectMemberRole()))){
+                continue;
+            }
+            userProjects.add(userProject);
+        }
+        return projectMapper.mapToProjectParticipantDtos(userProjects);
+    }
+
+    @Operation(description = "Создание проекта")
     @PostMapping
-    public ResponseEntity<CreatedResponse> createProject(@RequestBody ProjectCreateRequest projectCreateRequest) {
-        Organization organization = organizationService.getById(projectCreateRequest.ownerId());
+    public ResponseEntity<CreatedResponse> create(@RequestBody ProjectCreateRequest projectCreateRequest) throws UserNotFoundException {
+        User user = authService.getAuthenticatedUser();
+        if(projectCreateRequest.organizationId() == null){
+            if(projectService.isProjectExist(projectCreateRequest.name(), user)){
+                throw new RuntimeException("409 PROJECT_NAME_CONFLICT");
+            }
+            Project project = projectService.create(projectCreateRequest, user);
+            return ResponseEntity.status(201).body(new CreatedResponse(project.getId()));
+        }
+
+        Organization organization = organizationService.getById(projectCreateRequest.organizationId());
         if(projectService.isProjectExist(projectCreateRequest.name(), organization)){
-            //TODO: Реализовать корректную обработку ошибок
             throw new RuntimeException("409 PROJECT_NAME_CONFLICT");
         }
-        Project project = projectService.create(projectCreateRequest, organization);
+        if(organizationService.isUserAdminInOrganization(user, organization)){
+            throw new RuntimeException("403 FORBIDDEN_ORGANIZATION");
+        }
+        Project project = projectService.create(projectCreateRequest, user, organization);
         return ResponseEntity.status(201).body(new CreatedResponse(project.getId()));
     }
 
+    @Operation(description = "Получение проекта по id")
     @GetMapping("/{projectId}")
     public ProjectDetailsDto getProjectById(@PathVariable("projectId") Long projectId) throws UserNotFoundException {
         Project project = projectService.getById(projectId);
         User user = authService.getAuthenticatedUser();
         UserProject userProject = projectService.getUserProject(user, project);
+
+        if(project.getIsPrivate() && !projectService.isUserInProject(project, user)){
+            throw new RuntimeException("403 FORBIDDEN_PROJECT");
+        }
+
+        Organization organization = project.getOrganization();
+        if(organization != null && !organizationService.isOrganizationContainsUser(organization, user)){
+            throw new RuntimeException("403 FORBIDDEN_ORGANIZATION");
+        }
+
+        if(!project.getIsPrivate()
+                && organization != null
+                && !organizationService.isOrganizationContainsUser(organization, user)){
+            return projectMapper.mapToProjectDetailsDto(project);
+        }
+
         return projectMapper.mapToProjectDetailsDto(project, userProject);
     }
 
+    @Operation(description = "Обновление проекта по id")
     @PatchMapping("/{projectId}")
     public ProjectDetailsDto updateProjectById(
             @PathVariable("projectId") Long projectId, @RequestBody ProjectUpdateRequest request) throws UserNotFoundException {
         Project project = projectService.getById(projectId);
-        Project updatedProject = projectService.update(request, project);
-
         User user = authService.getAuthenticatedUser();
+        if(!projectService.isUserOwnerInProject(project, user)){
+            throw new RuntimeException("403 FORBIDDEN_PROJECT");
+        }
+
+        Project updatedProject = projectService.update(request, project);
         UserProject userProject = projectService.getUserProject(user, updatedProject);
-        return projectMapper.mapToProjectDetailsDto(updatedProject, userProject);
+
+        return projectMapper.mapToProjectDetailsDto(project, userProject);
     }
 
     @Operation(description = "Удаление проекта по id")
@@ -99,7 +176,7 @@ public class ProjectController {
         User user = authService.getAuthenticatedUser();
         Project project = projectService.getById(projectId);
 
-        if(!projectService.isOwner(project, user)){
+        if(!projectService.isUserOwnerInProject(project, user)){
             throw new RuntimeException("403 FORBIDDEN_PROJECT");
         }
 
@@ -108,28 +185,23 @@ public class ProjectController {
         return new DeletedResponse(true);
     }
 
+    @Operation(description = "Выход из проекта")
     @PostMapping("/{projectId}/leave")
     public LeftResponse leaveProjectById(@PathVariable("projectId") Long projectId) throws UserNotFoundException {
         User user = authService.getAuthenticatedUser();
         Project project = projectService.getById(projectId);
+        if(projectService.isUserOwnerInProject(project, user)){
+            throw new RuntimeException("400 OWNER CAN'T LEAVE PROJECT");
+        }
+        if(!projectService.isUserInProject(project, user)){
+            throw new RuntimeException("Пользователь не состоит в проекте");
+        }
         projectService.removeUserFromProject(user, project);
         return new LeftResponse(true);
     }
 
     //FIXME: Пагинация
-    @GetMapping("/{projectId}/participants")
-    public List<ProjectParticipantDto> getParticipants(@PathVariable("projectId") Long projectId){
-        Project project = projectService.getById(projectId);
-        Set<UserProject> userProjects = project.getUsers();
-        List<ProjectParticipantDto> participantDtos = new ArrayList<>();
-        for (UserProject userProject : userProjects) {
-            ProjectParticipantDto projectParticipantDto = projectMapper.mapToProjectParticipantDto(userProject);
-            participantDtos.add(projectParticipantDto);
-        }
-        return participantDtos;
-    }
-
-    //FIXME: Пагинация
+    @Operation(description = "Получение задач проекта по id")
     @GetMapping("/{projectId}/tasks")
     public List<ProjectTaskDto> getProjectTasks(@PathVariable("projectId") Long projectId){
         Project project = projectService.getById(projectId);
@@ -142,6 +214,7 @@ public class ProjectController {
         return projectTaskDtos;
     }
 
+    @Operation(description = "Создание задачи для проекта")
     @PostMapping("/{projectId}/tasks")
     public ResponseEntity<CreatedResponse> createTask(
             @PathVariable("projectId") Long projectId, @RequestBody TaskCreateRequest request) throws UserNotFoundException {
@@ -150,6 +223,7 @@ public class ProjectController {
         return ResponseEntity.status(201).body(new CreatedResponse(task.getId()));
     }
 
+    @Operation(description = "Получение задачи проекта")
     @GetMapping("/{projectId}/tasks/{taskId}")
     public ProjectTaskDto getTaskById(
             @PathVariable("projectId") Long projectId, @PathVariable("taskId") Long taskId
@@ -158,15 +232,48 @@ public class ProjectController {
         return taskMapper.mapToProjectTaskDto(task);
     }
 
-//    @PostMapping("/{projectId}/invites")
-//    public InviteResponse createInviteLink(
-//            @PathVariable("projectId") Long projectId, @RequestBody InviteRequest request){
-//
-//    }
+    @PostMapping("/{projectId}/invites")
+    public ProjectInviteResponse createInviteLink(
+            @PathVariable("projectId") Long projectId, @RequestBody ProjectInviteRequest request) throws UserNotFoundException {
+        Project project = projectService.getById(projectId);
+        User user = authService.getAuthenticatedUser();
+        if(!projectService.isUserOwnerInProject(project, user)){
+            throw new RuntimeException("403 FORBIDDEN_PROJECT");
+        }
+
+        ProjectInvite projectInvite = projectInviteService.create(request, project);
+        return projectInviteMapper.mapToProjectInviteResponse(projectInvite);
+    }
+
+    @PostMapping("/{projectId}/join")
+    public ProjectJoinedResponse join(@PathVariable("projectId") Long projectId) throws UserNotFoundException {
+        Project project = projectService.getById(projectId);
+        User user = authService.getAuthenticatedUser();
+        if(Boolean.FALSE.equals(project.getIsPrivate())){
+            throw new RuntimeException("403 PROJECT_NOT_PUBLIC");
+        }
+        if(projectService.isUserInProject(project, user)){
+            throw new RuntimeException("409 ALREADY_MEMBER");
+        }
+        if(project.getOrganization() == null){
+            throw new RuntimeException("403 FORBIDDEN_ORGANIZATION");
+        }
+        projectService.addUserToProject(user, project, ProjectMemberRole.MEMBER);
+        return new ProjectJoinedResponse(true, project.getId());
+    }
 
     @GetMapping("/public/search")
-    public List<ProjectListItemDto> getPublicProjects(){
-        List<Project> publicProjects = projectService.getPublicProjects();
-        return projectMapper.mapToProjectListItemDtos(publicProjects);
+    public List<ProjectListItemDto> getPublicProjects() throws UserNotFoundException {
+        User user = authService.getAuthenticatedUser();
+        List<Project> projects = projectService.getPublicProjects();
+        List<ProjectListItemDto> projectListItemDtos = new ArrayList<>();
+        for(Project project : projects){
+            if(projectService.isUserMemberOfProject(project, user)){
+                continue;
+            }
+            ProjectListItemDto projectListItemDto = projectMapper.mapToProjectListItemDto(project);
+            projectListItemDtos.add(projectListItemDto);
+        }
+        return projectListItemDtos;
     }
 }
